requirements.txt
openai
python-dotenv
playwright
azure-identity